--- components/AddButton/index.tsx
@@ -1,14 +1,0 @@
-import { Text, TouchableOpacity, TouchableOpacityProps } from "react-native";
-import { styles } from './style';
-
-type Props = TouchableOpacityProps & {
-    title: string
-}
-
-export function AddButton( {title, ...rest} : Props){
-    return(
-        <TouchableOpacity activeOpacity={0.5} style={styles.button} {...rest}>
-            <Text style={styles.title}>{title}</Text>
-        </TouchableOpacity>
-    )
-}
