--- components/AddButton/style.ts
@@ -1,17 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    button: {
-        width: "25%",
-        height: 45,
-        backgroundColor: "blue",
-        borderRadius: 7,
-        justifyContent: "center",
-        alignItems: "center"
-    },
-    title: {
-        fontSize: 16,
-        fontWeight: "bold",
-        color: "#FFFFFF"
-    }
-})
