--- components/CheckButton/style.ts
@@ -1,25 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    button1: {
-        width: "25%",
-        height: 45,
-        backgroundColor: "red",
-        borderRadius: 7,
-        justifyContent: "center",
-        alignItems: "center"
-    },
-    button2: {
-        width: "25%",
-        height: 45,
-        backgroundColor: "green",
-        borderRadius: 7,
-        justifyContent: "center",
-        alignItems: "center"
-    },
-    title: {
-        fontSize: 16,
-        fontWeight: "bold",
-        color: "#FFFFFF"
-    }
-})
