--- teste.txt
@@ -1,1 +1,0 @@
-teste de conexão
